+++ pkg/query.go
@@ -6,6 +6,7 @@ import (
 	"encoding/json"
 	"fmt"
 	_data "github.com/michelin/snowflake-grafana-datasource/pkg/data"
+	"github.com/michelin/snowflake-grafana-datasource/pkg/query"
 	"github.com/michelin/snowflake-grafana-datasource/pkg/utils"
 	"math/big"
 	"reflect"
@@ -27,13 +28,6 @@ var float float64
 var str string
 var integer int64
 
-// Constant used to describe the time series fill mode if no value has been seen
-const (
-	NullFill     = "null"
-	PreviousFill = "previous"
-	ValueFill    = "value"
-)
-
 type queryModel struct {
 	QueryText   string   `json:"queryText"`
 	QueryType   string   `json:"queryType"`
@@ -41,8 +35,8 @@ type queryModel struct {
 	FillMode    string   `json:"fillMode"`
 }
 
-func fetchData(ctx context.Context, qc *_data.QueryConfigStruct, config *pluginConfig, password string, privateKey string) (result _data.QueryResult, err error) {
-	connectionString := getConnectionString(config, password, privateKey)
+func fetchData(ctx context.Context, qc *_data.QueryConfigStruct, config *pluginConfig, authenticationSecret _data.AuthenticationSecret) (result _data.QueryResult, err error) {
+	connectionString := getConnectionString(config, authenticationSecret)
 
 	db, err := sql.Open("snowflake", connectionString)
 	if err != nil {
@@ -164,7 +158,7 @@ func transformQueryResult(qc _data.QueryConfigStruct, columnTypes []*sql.ColumnT
 	return values, nil
 }
 
-func (td *SnowflakeDatasource) query(ctx context.Context, dataQuery backend.DataQuery, request *backend.QueryDataRequest, config pluginConfig, password string, privateKey string) (response backend.DataResponse) {
+func (td *SnowflakeDatasource) query(ctx context.Context, dataQuery backend.DataQuery, request *backend.QueryDataRequest, config pluginConfig, authentication _data.AuthenticationSecret) (response backend.DataResponse) {
 	var qm queryModel
 	err := json.Unmarshal(dataQuery.JSON, &qm)
 	if err != nil {
@@ -195,7 +189,7 @@ func (td *SnowflakeDatasource) query(ctx context.Context, dataQuery backend.Data
 	log.DefaultLogger.Info("Query config", "config", qm)
 
 	// Apply macros
-	queryConfig.FinalQuery, err = Interpolate(&queryConfig)
+	queryConfig.FinalQuery, err = query.Interpolate(&queryConfig)
 	if err != nil {
 		response.Error = err
 		return response
@@ -205,7 +199,7 @@ func (td *SnowflakeDatasource) query(ctx context.Context, dataQuery backend.Data
 	queryConfig.FinalQuery = strings.TrimSuffix(strings.TrimSpace(queryConfig.FinalQuery), ";")
 
 	frame := data.NewFrame("")
-	dataResponse, err := fetchData(ctx, &queryConfig, &config, password, privateKey)
+	dataResponse, err := fetchData(ctx, &queryConfig, &config, authentication)
 	if err != nil {
 		response.Error = err
 		return response
@@ -214,9 +208,6 @@ func (td *SnowflakeDatasource) query(ctx context.Context, dataQuery backend.Data
 	for _, table := range dataResponse.Tables {
 		timeColumnIndex := -1
 		for i, column := range table.Columns {
-			if err != nil {
-				return backend.DataResponse{}
-			}
 			// Check time column
 			if queryConfig.IsTimeSeriesType() && utils.EqualsIgnoreCase(queryConfig.TimeColumns, column.Name()) {
 				if strings.EqualFold(column.Name(), "Time") {
@@ -287,7 +278,7 @@ func (td *SnowflakeDatasource) query(ctx context.Context, dataQuery backend.Data
 func (td *SnowflakeDatasource) longToWide(frame *data.Frame, queryConfig _data.QueryConfigStruct, dataResponse _data.QueryResult) (*data.Frame, error) {
 	tsSchema := frame.TimeSeriesSchema()
 	if tsSchema.Type == data.TimeSeriesTypeLong {
-		fillMode := &data.FillMissing{Mode: mapFillMode(queryConfig.FillMode), Value: queryConfig.FillValue}
+		fillMode := &data.FillMissing{Mode: query.MapFillMode(queryConfig.FillMode), Value: queryConfig.FillValue}
 		if len(dataResponse.Tables) > 0 && len(dataResponse.Tables[0].Rows) > 0 {
 			var err error
 			frame, err = data.LongToWide(frame, fillMode)
@@ -307,21 +298,6 @@ func (td *SnowflakeDatasource) longToWide(frame *data.Frame, queryConfig _data.Q
 	return frame, nil
 }
 
-func mapFillMode(fillModeString string) data.FillMode {
-	var fillMode = data.FillModeNull
-	switch fillModeString {
-	case ValueFill:
-		fillMode = data.FillModeValue
-	case NullFill:
-		fillMode = data.FillModeNull
-	case PreviousFill:
-		fillMode = data.FillModePrevious
-	default:
-		// no-op
-	}
-	return fillMode
-}
-
 func fillTimesSeries(queryConfig _data.QueryConfigStruct, intervalStart int64, intervalEnd int64, timeColumnIndex int, frame *data.Frame, columnSize int, count *int, previousRow []interface{}) {
 	if queryConfig.IsTimeSeriesType() && queryConfig.FillMode != "" && timeColumnIndex != -1 {
 		for stepTime := intervalStart + queryConfig.Interval.Nanoseconds()/1e6*int64(*count); stepTime < intervalEnd; stepTime = stepTime + (queryConfig.Interval.Nanoseconds() / 1e6) {
@@ -332,11 +308,11 @@ func fillTimesSeries(queryConfig _data.QueryConfigStruct, intervalStart int64, i
 					continue
 				}
 				switch queryConfig.FillMode {
-				case ValueFill:
+				case query.ValueFill:
 					frame.Fields[i].Append(&queryConfig.FillValue)
-				case NullFill:
+				case query.NullFill:
 					frame.Fields[i].Append(nil)
-				case PreviousFill:
+				case query.PreviousFill:
 					if previousRow == nil {
 						utils.InsertFrameField(frame, nil, i)
 					} else {